aggregator/requirements.txt
pygresql
python-dotenv
